New Brunswick Tuition Amounts Fixes #292
